Refactor merchant tests to have files dedicated to each endpoint
